Moved Street normalizers to subdirectory
