+++ src/Normalizer/FromJson/Street/StreetNameDetailNormalizer.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson;
+namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Street;
 
 use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Geographical\GeographicalNamesNormalizer;
+use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\IdExtractor;
 use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Locality\LocalityNameNormalizer;
 use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameDetail;
 use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId;
@@ -23,11 +24,12 @@ final class StreetNameDetailNormalizer
      */
     public function normalize(object $jsonData): StreetNameDetail
     {
+        $idExtractor = new IdExtractor();
         $geoGraphicalNamesNormalizer = new GeographicalNamesNormalizer();
         $localityNameNormalizer = new LocalityNameNormalizer();
 
         return new StreetNameDetail(
-            new StreetNameId((int) $jsonData->identificator->objectId),
+            new StreetNameId($idExtractor->extractObjectId($jsonData)),
             $geoGraphicalNamesNormalizer->normalize($jsonData->straatnamen),
             $localityNameNormalizer->normalize($jsonData->gemeente)
         );
