+++ src/Normalizer/FromJson/Street/StreetNamesNormalizer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson;
+namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Street;
 
 use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNames;
