+++ tests/Normalize/FromJson/Street/StreetNamesNormalizerTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Tests\Flanders\BasicRegisters\Normalize\FromJson;
+namespace DigipolisGent\Tests\Flanders\BasicRegisters\Normalize\FromJson\Street;
 
-use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\StreetNamesNormalizer;
+use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Street\StreetNamesNormalizer;
 use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetName;
@@ -13,7 +13,7 @@ use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNames;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\StreetNamesNormalizer
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Street\StreetNamesNormalizer
  */
 class StreetNamesNormalizerTest extends TestCase
 {
